adk_agents/pyproject.toml
[project]
name = "adk-agents"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-adk>=0.5.0",
    "pandas>=2.2.3",
]
